Database setup scripts are more robust
The scripts rely on comparing the exact output of `psql` commands to
check for the existence of objects in the database.

The psql output can be impacted by user preferences in their `.psqlrc`
file, breaking the scripts.

To reproduce, add the following line to `~/.psqlrc`:

```
\pset format wrapped
```

The install/uninstall scripts will now fail to find existing objects.

The -q flag silences all output other than the query result.
The -X flag ignores the user's .psqlrc so that output is more deterministic.