Remove redundant indexes
There are already multicolumn indexes on stream,position and
category/global_position. Those cover the two most common queries.

PostgreSQL can still use those indexes for queries on stream or category
alone, since they are the leftmost column:

> A multicolumn B-tree index can be used with query conditions that
involve any subset of the index's columns, but the index is most
efficient when there are constraints on the leading (leftmost) columns.
The exact rule is that equality constraints on leading columns, plus any
inequality constraints on the first column that does not have an
equality constraint, will be used to limit the portion of the index that
is scanned

https://www.postgresql.org/docs/9.6/static/indexes-multicolumn.html

The redundant single column indexes may have a slight performance
advantage over the multi-column indexes for queries that only use the
first column (since the indexes are smaller), but those queries are
likely rare, and probably not worth the cost of the redundant indexes
(disk space, and slower INSERTs).